Create install.sh
I'm sure this file needs testing and a little more love I'm sure, but this will install 1.2.8 (as it is today) and 2.0-pre on a fresh install of both Ubuntu 14+ and CentOS 7+.

The script should detect which distro it's running on and then offers the user several questions to install Snipe-IT with.  I understand security is a huge thing, so as part of an "automatic" install, I tell the user if they choose to allow the script to generate mysql passwords, that they should record then blow the file away (or the script can do it later on.) 

chmod +x install.sh
sudo ./install.sh

Thanks!