Create install.sh
This script will

1. Ask for Snipe it version 1.2.8 or 2.0
2. Detect distro version (between CentOS and Debian/Ubuntu)
3. Install all software and dependencies, create vhosts.

This isn't the smartest script, and it needs a little love, and I'm on the fence about SMTP settings.  Would love feedback about the SMTP part, and the whole script in general.  I'll be glad to fine tune it some more,  I just need a point in the right direction. Hopefully if this is merged in 2.0, it won't cause a huge influx of tickets.  This is the first time I've contributed to a project like this, so I don't know the social norms; if having my name at the top isn't appropriate (or fit in with the scope of Snipe It), we can remove it. No big deal, just wanted to have my name somewhere, but it's not a make-it-or-break-it thing for me :o)