Install Snipe-It on a clean Ubuntu/CentOS install
This script is designed to install Snipe-It on a fresh install of Ubuntu or CentOS.  You will receive minimal options regarding setting passwords and the script should take care of the rest.  The purpose of this script is to install Snipe-It on a server that will be dedicated to the application.

This was tested on Ubuntu 14.04 and CentOS 7, YMMV with different OS versions.