Issue #2535 - Update Product Key Field DB Entry
Updates the product key field from ~255< to a varchar entry in the database with 2048 character entry.

This should appropriately fix the limited entry for serial numbers.

Might want to double check my laravel. Still trying to glass over the code, but, if I understood it right, this is the location of where it needs to be changed?