Improve German translations
This commit changes some of the German translations to improve compatibility with the English ones. For example, "address" in English is (not) translated with "Supplier Address", which is apparently wrong.

Some missing translations are added, and some minor grammar mistakes are fixed.

Certain phrases are adapted to achieve a more consistent nomenclature and style.

Previously, formal and informal forms of address were used inconsistently. Now, the formal salutation ("Sie") is used.