Clean up old javascript
This tries to move as much javsacript as possible to
node_modules/versioned packages. It deletes a number of packages stored
in resources/assets and public/js/plugins.  It also removes the
bower_components directory that we don't really use any more.  There are
a few plugins (like iCheck) that we keep in public because of the images
they reference, and how unhappy webpack is with handling those images
from node_modules.  Has received some testing, but could certainly do
with more.

More Progress.

Progress.