Don't reverse the sort order, it's confusing.
The API reverses the activity sort order. This is confusing and causes
problems for things like the recenty activity on the dashboard - to get
things in in descending order it needed to specify ascending order. Making
this change makes the dashboard behave as expected.

Possibly the activity report needs its default sort order setting too?