Fix print label formatting
There were a couple of margin issues. We need to subtract the
horizontal guttering from the right margin, and the vertical
guttering from the bottom margin. There was also an issue where
inline-block's add whitespace between the elements when there's
whitespace in the HTML. The solution is to set the font-size to 0 on
the parent element, and back to the correct size on the labels.

Also fixed an issue whereby it added a blank page if your page had
exactly the number of assets as the page limit.