API Route's file cleanup.
We can add a prefix to all API routes in the "Route::Group" of api,
which allows the names below to be much less verbose.  In addition, it
allows to avoid overriding the names parameters in each Route::resource.

Also, by moving to Route::apiResource instead of Route::resource(), it
gives us omission of create and edit routes for free.