Fix #4275: advanced search using custom fields
When filtering assets using custom fields, the bootstrap table
toolbar tries to match its header columns with the properties
in the root of the item object.

Since the custom fields are an array inside the property
`custom_items` of the item instead of flatten properties, the check
done by bootstrap table will always be false, resulting in the
exclusion of all items searched using a custom field.