Fix advanced search with serial and another field
The advanced search in /hardware produces incorrect results if the serial is combined with another field like category.
The query uses an OR-Clause to combine the fields.

This is because the filedname in Assets::scopeByFilter is incorrectly named as 'product_key', not serial.
The serial-field gets caught by the last generic If-Statement and the OR-Clause gets added:
	$query->orWhere('assets.'.$fieldname, 'LIKE', '%' . $search_val . '%');

Fix this by changing the fieldname to 'serial'.