Preserve backward compatibility for locations with companies
With FullMultipleCompanySupport enabled, a standard user can only see locations from its own company.
This is a backward incompatible change. To preserve backward compatibility, locations with company_id = null are shown to all users.

Note that if a scoped user with permissions to change locations updates a location, the location inherits the company of the user.