Change the way userDn is build for non-AD login
Currently, snipe-it relies on a manually crafted user DN to bind a user
and therefore check if they exist, and *then*, it searches in LDAP if
the user is found when applying the filter.

This works nicely when the base_dn is "fixed". In a company with
multiple business lines or business units, though, there might be more
than one OU for LDAP users, and in that case, crafting the DN manually
might fail.

This patch aims at crafting the userDn from a ldap_search, assuming the
user field is unique. If more than 1 user is found, the user field is
not unique and the login method will throw.

This increases significantly the LDAP login flexibility.

Fixes #12331